Refactored again for SAxPRO backport. SAxPRO code isn't working yet!
