+++ Firmware/CoverUI/YardForce/include/WidgetTextTicker.hpp
