+++ Firmware/CoverUI/YardForce/include/model_RM-ECOW-V100.h
@@ -12,10 +12,14 @@
 #define YARDFORCE_MDL_RMECOWV100_H
 
 #define MDL_RMECOWV100
-#define LEDCTRL_HDR "YardForce/LEDcontrol_RM-ECOW-V100.hpp"
-#define BUTTONS_HDR "YardForce/buttons_RM-ECOW-V100.hpp"
+
+#define HAS_LEDS
+#define LEDCTRL_HDR "YardForce/include/model/LEDcontrol_RM-ECOW-V100.hpp"
+
+#define BUTTONS_HDR "YardForce/include/model/buttons_RM-ECOW-V100.hpp"
+
 #ifdef MOD_EMERGENCY
-#define EMERGENCY_HDR "YardForce/emergency_RM-ECOW-V100.hpp"
+#define EMERGENCY_HDR "YardForce/include/model/emergency_RM-ECOW-V100.hpp"
 #endif
 
 // LowLevel Serial Pins
