+++ Firmware/CoverUI/platformio.ini
@@ -41,13 +41,14 @@ build_src_filter =
   -<LEDcontrol.cpp>
   -<statemachine.cpp>
   -<build/**>
-  +<YardForce/*>
+  +<YardForce/**>
   +<YardForce/*.hpp>
+  -<YardForce/include/model/**>
+  -<YardForce/include/UC1698*>
+  -<YardForce/include/Widget*>
+  -<YardForce/include/WYM240128K1*>
   -<YardForce/old/**>
   -<YardForce/assets/**>
-  -<YardForce/UC1698.*>
-  -<YardForce/WYM240128K1.*>
-  -<YardForce/Widget*>
 
 build_flags =
   -DHW_YF
@@ -68,7 +69,7 @@ board = disco_f030r8
 build_flags =
   ${env.build_flags}
   -DMCU_STM32
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
 
 upload_flags =
   -f
@@ -91,7 +92,7 @@ build_flags =
   ${env.build_flags}
   -DMCU_STM32
   -DMOD_RAIN
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
 
 upload_flags =
   -f
@@ -115,7 +116,7 @@ build_flags =
   -DMCU_STM32
   -DMOD_RAIN
   -DMOD_EMERGENCY
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
 
 upload_flags =
   -f
@@ -138,7 +139,7 @@ build_flags =
   ${env.build_flags}
   -DMCU_STM32
   -DMOD_EMERGENCY
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
 
 upload_flags =
   -f
@@ -162,7 +163,7 @@ build_flags =
   -DMCU_STM32
   -DMOD_RAIN_OFF
   -DMOD_EMERGENCY_OFF
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
 
 ; Need at least openocd 0.11 for working CMSIS-DAP adapter detection
 platform_packages = platformio/tool-openocd@^2.1100.211028
@@ -199,7 +200,7 @@ board = genericGD32F330R8
 build_flags =
   ${env.build_flags}
   -DMCU_GD32
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
   
 upload_flags =
   -f
@@ -226,7 +227,7 @@ build_flags =
   ${env.build_flags}
   -DMCU_GD32
   -DMOD_RAIN
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
 
 upload_flags =
   -f
@@ -253,7 +254,7 @@ build_flags =
   -DMCU_GD32
   -DMOD_RAIN
   -DMOD_EMERGENCY
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
 
 upload_flags =
   -f
@@ -280,7 +281,7 @@ build_flags =
   ${env.build_flags}
   -DMCU_GD32
   -DMOD_EMERGENCY
-  -include YardForce/model_C500.h
+  -include YardForce/include/model_C500.h
 
 upload_flags =
   -f
@@ -305,7 +306,7 @@ board = disco_f030r8
 build_flags =
   ${env.build_flags}
   -DMCU_STM32
-  -include YardForce/model_RM-ECOW-V100.h
+  -include YardForce/include/model_RM-ECOW-V100.h
 
 upload_flags =
   -f
@@ -327,7 +328,7 @@ build_flags =
   ${env.build_flags}
   -DMCU_STM32
   -DMOD_EMERGENCY
-  -include YardForce/model_RM-ECOW-V100.h
+  -include YardForce/include/model_RM-ECOW-V100.h
 
 upload_flags =
   -f
@@ -348,9 +349,6 @@ debug_server =
 ;*   YardForce SAxPRO, STM32 Variants   *
 ;********************************************
 
-; TODO:
-; - UC1698?? & Display Class doesn't need to be compiled by the other envs
-
 [env:YF_SAXPRO_STM32]
 board = disco_f030r8
 board_upload.maximum_ram_size = 32768 ; STM32F030RCx has 32k SRAM
@@ -358,20 +356,17 @@ board_upload.maximum_size = 262144    ; STM32F030RCx has 256k Flash
 
 build_src_filter =
   ${env.build_src_filter}
-  -<YardForce/Rain.*>
-  -<YardForce/Emergency.*>
+  -<YardForce/include/Rain.*>
+  -<YardForce/include/Emergency.*>
   +<YardForce/assets/*>
-  +<YardForce/UC1698.*>
-  +<YardForce/WYM240128K1.*>
-  +<YardForce/Widget*>
 
 build_flags =
   ${env.build_flags}
-  -DMDL_SAXPRO
   -DMCU_STM32
   -DLV_CONF_INCLUDE_SIMPLE
   -DLV_LVGL_H_INCLUDE_SIMPLE
   -IYardForce ; Required by LVGL to find his lv_conf.h
+  -include YardForce/include/model_SAxPRO.h
 
 lib_deps = lvgl/lvgl@^8.3.9
